+++ src/app/settings/llm/Ollama/index.tsx
@@ -1,5 +1,4 @@
 import { Ollama } from '@lobehub/icons';
-import { useTheme } from 'antd-style';
 import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -10,7 +9,6 @@ import Checker from './Checker';
 
 const OllamaProvider = memo(() => {
   const { t } = useTranslation('setting');
-  const theme = useTheme();
 
   return (
     <ProviderConfig
@@ -20,13 +18,12 @@ const OllamaProvider = memo(() => {
         label: t('llm.checker.title'),
         minWidth: undefined,
       }}
-      modelList={{ showModelFetcher: true }}
       provider={ModelProvider.Ollama}
       showApiKey={false}
+      showBrowserRequest
       showEndpoint
-      title={
-        <Ollama.Combine color={theme.isDarkMode ? theme.colorText : theme.colorPrimary} size={24} />
-      }
+      title={<Ollama.Combine size={24} />}
+      // modelList={{ showModelFetcher: true }}
     />
   );
 });